Node deploy cmd (#1154)
* almost

* add node deploy command

* mv find ledger indices to helpers

* fix endpoint issue

* lint

* avoid prompt for devnet endpoint on create

* nit

* move helpers only constants to helpers

* change to node devnet deploy

* address PR comments

* lint
