+++ cmd/nodecmd/status.go
@@ -74,11 +74,11 @@ func statusNode(_ *cobra.Command, args []string) error {
 			return err
 		}
 	}
-	notHealthyNodes, err := checkClusterIsHealthy(clusterName)
+	notBootstrappedNodes, err := checkClusterIsBootstrapped(clusterName)
 	if err != nil {
 		return err
 	}
-	notBootstrappedNodes, err := checkClusterIsBootstrapped(clusterName)
+	notHealthyNodes, err := checkClusterIsHealthy(clusterName)
 	if err != nil {
 		return err
 	}
