+++ cmd/nodecmd/sync.go
@@ -138,6 +138,13 @@ func syncSubnet(_ *cobra.Command, args []string) error {
 	if len(notBootstrappedNodes) > 0 {
 		return fmt.Errorf("node(s) %s are not bootstrapped yet, please try again later", notBootstrappedNodes)
 	}
+	notHealthyNodes, err := checkClusterIsHealthy(clusterName)
+	if err != nil {
+		return err
+	}
+	if len(notHealthyNodes) > 0 {
+		return fmt.Errorf("node(s) %s are not healthy, please fix the issue and again", notHealthyNodes)
+	}
 	incompatibleNodes, err := checkAvalancheGoVersionCompatible(clusterName, subnetName)
 	if err != nil {
 		return err
