+++ cmd/subnetcmd/addValidator.go
@@ -131,7 +131,7 @@ func addValidator(_ *cobra.Command, args []string) error {
 	}
 
 	// get keychain accesor
-	kc, err := GetKeychain(useLedger, ledgerAddresses, keyName, network)
+	kc, err := GetKeychain(false, useLedger, ledgerAddresses, keyName, network)
 	if err != nil {
 		return err
 	}
