+++ cmd/subnetcmd/elastic.go
@@ -303,7 +303,7 @@ func transformElasticSubnet(cmd *cobra.Command, args []string) error {
 	}
 
 	// get keychain accessor
-	kc, err := GetKeychain(useLedger, ledgerAddresses, keyName, network)
+	kc, err := GetKeychain(false, useLedger, ledgerAddresses, keyName, network)
 	if err != nil {
 		return err
 	}
